+++ solution/RenderEngine/include/PrimitiveMeshFactory.h
@@ -18,13 +18,13 @@ public:
 	GUIMeshData* createPlaneBottomUp(const glm::vec2& size);
 	MeshData* createPlane(float size, float tileMultiplier);
 	MeshData* createPlane(const glm::vec2& size, float tileMultiplier);
-	void createPlaneIndices(std::vector<uint16_t> &result, int planesCount = 1);
+	void createPlaneIndices(std::vector<uint32_t> &result, int planesCount = 1);
 	MeshData* createCube(float size);
 	MeshData* createSphere(float radius = 1.0f, uint16_t stackCount = 16, uint16_t sectorCount = 32);
 
 private:
 	GUIMeshData* createPlane2D(const glm::vec2 &size, const float* texCoords);
-	void calculateTangentsAndBitangents(std::vector<VertexData>& vertexData, std::vector<uint16_t> indices, int facesCount);
+	void calculateTangentsAndBitangents(std::vector<VertexData>& vertexData, std::vector<uint32_t> indices, int facesCount);
 	template <typename T>
 	void getPlaneUVs(std::vector<T> &vertexData, const float *planeTexCoords)
 	{
@@ -35,7 +35,7 @@ private:
 		}
 	}
 
-	void createSphereIndices(std::vector<uint16_t>& indices, uint16_t stackCount, uint16_t sectorCount);
+	void createSphereIndices(std::vector<uint32_t>& indices, uint16_t stackCount, uint16_t sectorCount);
 };
 
 } // namespace
