+++ solution/RenderEngine/src/core/wrappers/graphics/OpenGLESAPI.cpp
@@ -49,7 +49,7 @@ void OpenGLESAPI::createEBO(MeshData* meshData)
 	int size = meshData->indices.size();
 
 	GL_CHECK(glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, meshData->ebo));
-	GL_CHECK(glBufferData(GL_ELEMENT_ARRAY_BUFFER, size * sizeof(uint16_t), &meshData->indices[0], GL_STATIC_DRAW));
+	GL_CHECK(glBufferData(GL_ELEMENT_ARRAY_BUFFER, size * sizeof(uint32_t), &meshData->indices[0], GL_STATIC_DRAW));
 }
 
 void OpenGLESAPI::createGUIVAO(GUIMeshData* meshData, uint32_t maxItems, bool isDynamic)
@@ -88,7 +88,7 @@ void OpenGLESAPI::createGUIEBO(GUIMeshData* meshData, uint32_t maxItems, bool is
 	// EBO
 	GL_CHECK(glGenBuffers(1, &meshData->ebo));
 	GL_CHECK(glBindBuffer(GL_ELEMENT_ARRAY_BUFFER, meshData->ebo));
-	GL_CHECK(glBufferData(GL_ELEMENT_ARRAY_BUFFER, size * sizeof(uint16_t), data, usage));
+	GL_CHECK(glBufferData(GL_ELEMENT_ARRAY_BUFFER, size * sizeof(uint32_t), data, usage));
 }
 
 void OpenGLESAPI::bindVAO(uint32_t vao, uint32_t vbo)
